moving param views to separate h/cpp file
